Update README.md
Hit me with some code snippets that y'all are proud of and we can get those included as well! This was the only thing that I did that I was proud of, and it omits the fact that I bungled delete! Also feel free to make any edits to the members and responsibilities if this wasn't flattering enough! I wasn't super close to a lot of what y'all were doing for the first day or so, so please do not be mad at me if this does not do you justice! Just edit it and we can show each other love for a job well done IRL! =)